+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bug fixes
  * [[`PR-70`](https://github.com/thiagoesteves/deployex/pull/70)] Changed function listener to subscribe
  * [[`PR-71`](https://github.com/thiagoesteves/deployex/pull/71)] Added ex_docs and enhanced documentation
+ * [[`PR-72`](https://github.com/thiagoesteves/deployex/pull/72)] Change deploy reference to string instead erlang reference
 
 ## 0.3.0-rc20 (2024-10-02)